Add support for %20 encoded URLs
Because URI.parse cannot handle whitespace the URLs need to be encoded with %20. This
then causes problems on the Browse Everything end since it assumes that the path is
valid as is. This patch tries to escape any special HTML characters on the way out so
that download works as expected