Show the full title of the file when hovering over its link
Listing of files are truncated so they do not push other columns out the modal's width.
However, in some cases, you need to see the full name of the file if you have several
long filenames that each begin with the same text.

Providing a "title" attribute to the filename's row uses the browser's hover feature
so that the full length of the filename can be seen.