Easing version restriction of Thor
Prior to this commit, the version restriction of Thor was locked to a
version compatable the various Rails versions, but reflective of the
somewhat implicit fixing pre 1.0 gem dependencies to a narrow version.

With the release of Thor 1.0 and Rails adherance to these ranges, this
commit opts for adding an explicit version range. _Note: We could
remove the explicit dependency, and allow for the upstream Rails
dependency to take effect._

Closes #338