Ignoring .ruby-version file
The .ruby-version is something that is often in place for rbenv
installations of Ruby. For a gem's development, this is not something
that I feel is important to have as part of the repository. It is,
however, something useful for local development. So, instead of
imposing my .ruby-version on everyone else, I'm instead opting to
set this as an ignored file.