Remove karma/jasmine JS tests, and all infrastructure
These are not working and have not been working for some time -- unclear how long. There was only one very simple/unclear-value test in there anyway.

Remove the whole karma/jasmine infrastructure, as it wasn't working, we don't currently have the capaccity to get it working, and it's one more moving part to go wrong.