Fix "no turbolinks" test to test something different
It does NOT really test "no turbolinks", as there will still be a `window.Turbolinks` object, and a `turbolinks:load` event, so
it's not really a no-turbolinks path. It actually will test as if the page we are navigating to was the FIRST page loaded in
the turbolinks session, which is still a separate path to test, sure.