If we're going to refer to `Rails` object, we need to import it
The code was referring to it to retrieive the CSRF token to inject. When manually running internal_test_app in dev mode, it failed to do this injection, so task failed. This fixed it.

The feature tests (once hacked to run) were green without this... which is a mystery to me, I don't know what is different in test environment that made this work without this change. But tests are still green after this too.