Exclude dummy app from rubocop
It was generated by a Rails generator, and a bunch of it out of the box is not acceptable to rubocop, we don't want to alter it, we want to keep what Rails genreates with as few changes as possible. Just have rubocop ignore the whole thing, even though that means ignoring future custom changes we make to it too, oh well.