FileSystem driver has new allow_relative_home param
If set to true, then `home` can be a relative path (relative to Rails.root).

The use case is mainly for CI/dummy test app setup. Added new param to make it opt-in, to make sure it's completely backwards compat and does not introduce any potential security issues -- you need to opt in to new supported behavior with new allow_relative_home param.