Add support for Sharepoint (Joined Teams and OneDrive)
Co-authored-by: Rory McNicholl <rory.mcnicholl@london.ac.uk>
Refer to https://github.com/samvera/browse-everything/pull/418 for
what is being built off of.

Refer to v1.5-avalon branch for full info on squashed commits:
- Implement authorization_code flow
- Update SharePoint.md
- Return all user accessible sharepoint sites
- Add an option to filter what sites get returned
- Use select param to limit returned metadata:
We do not use every metadata field returned from the all sites and all
drives queries. Using the select parameter to limit what gets returned
to the handful of fields we need should increase the speed of the
requests, resulting in faster loading.
- Update documentation
- Add tests for sharepoint driver
- Enable permissions re-consent flow
- Use joinedTeams as base folder instead of sites:
The sites endpoint returns every single sharepoint site that a user has
access to. This list can include many more options than a user would
expect, including stuff that the user should technically not have access
to (such as if a site has its permissions misconfigured). While we can
use a $search param to narrow the results, we can only get so granular
and so even the filtered results may provide an unwieldy amount of
folders. Because IU is interested in uploading files that are stored on
sharepoint drives that originate from Microsoft Teams, we use the
`joinedTeams` endpoint to only return drives that a user is explicitly a
member of. This creates a much more sensible and manageable list for
users to browse through and one that should have no surprises for the
user.
- Update SharePoint.md
- Initial Bootstrap 5 upgrade:
This commit is to update the gem enough to load in Avalon Media System
when running Bootstrap 5. Functionality and styling seem to be working
but it has not been thoroughly tested.
- Cleanup of documentation and code
- Remove unneeded methods